typescriptify unit test
