+++ modules/standards/test/encrypt-password-in-request-data.ts
@@ -1,13 +1,14 @@
-// @flow
-
 import { it, describe } from 'mocha'
-import powerCrypt from 'power-crypt/jsnext'
+// @ts-ignore
+import powerCrypt from 'power-crypt'
+// @ts-ignore
 import assert from 'power-assert'
-import { encryptPasswordInRequestData } from '../src/encrypt-password-in-request-data.js'
+import { GeneralRequestData } from '@phenyl/interfaces'
+import { encryptPasswordInRequestData } from '../src/encrypt-password-in-request-data'
 
-describe('encryptPasswordInRequestData', function () {
-  it ('does nothing if request isnt update', function () {
-    const requestData = {
+describe('encryptPasswordInRequestData', () => {
+  it ('does nothing if request isnt update', () => {
+    const requestData: GeneralRequestData = {
       method: 'get',
       payload: {
         entityName: 'user',
@@ -19,8 +20,8 @@ describe('encryptPasswordInRequestData', function () {
     assert.deepEqual(encryptedRequestData, requestData)
   })
 
-  it ('encrypts password if password is in request data with insertOne method', function () {
-    const requestData = {
+  it ('encrypts password if password is in request data with insertOne method', () => {
+    const requestData: GeneralRequestData = {
       method: 'insertOne',
       payload: {
         entityName: 'user',
@@ -41,8 +42,8 @@ describe('encryptPasswordInRequestData', function () {
     assert.deepEqual(encryptedRequestData, expectedRequestData)
   })
 
-  it ('encrypts password if password is in request data with insertMulti method', function () {
-    const requestData = {
+  it ('encrypts password if password is in request data with insertMulti method', () => {
+    const requestData: GeneralRequestData = {
       method: 'insertMulti',
       payload: {
         entityName: 'user',
@@ -63,8 +64,8 @@ describe('encryptPasswordInRequestData', function () {
     assert.deepEqual(encryptedRequestData, expectedRequestData)
   })
 
-  it ('encrypts password if password is in request data with update method', function () {
-    const requestData = {
+  it ('encrypts password if password is in request data with update method', () => {
+    const requestData: GeneralRequestData = {
       method: 'updateById',
       payload: {
         id: 'foo',
